requirements.txt
requests
python-daemon
pyyaml
websocket-client
slackclient
beautifulsoup4
wolframalpha
pymongo
